change float to double

+++ jsettlers.logic/src/main/java/jsettlers/ai/highlevel/AiBuildingConstants.java
@@ -18,22 +18,22 @@ package jsettlers.ai.highlevel;
  * @author codingberlin
  */
 public class AiBuildingConstants {
-	public static final double COAL_MINE_TO_IRONORE_MINE_RATIO = 2F / 1F;
-	public static final double WEAPON_SMITH_TO_BARRACKS_RATIO = 3F / 1F;
-	public static final double WEAPON_SMITH_TO_FISHER_HUT_RATIO = 2F / 3F;
-	public static final double WEAPON_SMITH_TO_FARM_RATIO = 2F / 1F;
-	public static final double WEAPON_SMITH_TO_IRONMELT_RATIO = 1F / 1F;
-	public static final double IRONMELT_TO_WEAPON_SMITH_RATIO = 1F / WEAPON_SMITH_TO_IRONMELT_RATIO;
-	public static final double FARM_TO_BAKER_RATIO = 1F / 1F;
-	public static final double FARM_TO_MILL_RATIO = 3F / 1F;
-	public static final double FARM_TO_WATERWORKS_RATIO = 3F / 1F;
-	public static final double FARM_TO_PIG_FARM_RATIO = 3F / 1F;
-	public static final double FARM_TO_SLAUGHTER_RATIO = 6F / 1F;
-	public static final double WEAPON_SMITH_TO_LUMBERJACK_RATIO = 20F / 8F;
-	public static final double LUMBERJACK_TO_SAWMILL_RATIO = 2F / 1F;
-	public static final double LUMBERJACK_TO_FORESTER_RATIO = 2F / 1F;
-	public static final double LUMBERJACK_TO_STONE_CUTTER_RATIO = 8F / 5F;
-	public static final double COAL_MINE_TO_SMITH_RATIO = 1F / 1.8F;
-	public static final double IRON_MINE_TO_IRONMELT_RATIO = COAL_MINE_TO_SMITH_RATIO / COAL_MINE_TO_IRONORE_MINE_RATIO;
-	public static final double WINEGROWER_TO_TEMPLE_RATIO = 1F / 1F;
+	public static final float COAL_MINE_TO_IRONORE_MINE_RATIO = 2F / 1F;
+	public static final float WEAPON_SMITH_TO_BARRACKS_RATIO = 3F / 1F;
+	public static final float WEAPON_SMITH_TO_FISHER_HUT_RATIO = 2F / 3F;
+	public static final float WEAPON_SMITH_TO_FARM_RATIO = 2F / 1F;
+	public static final float WEAPON_SMITH_TO_IRONMELT_RATIO = 1F / 1F;
+	public static final float IRONMELT_TO_WEAPON_SMITH_RATIO = 1F / WEAPON_SMITH_TO_IRONMELT_RATIO;
+	public static final float FARM_TO_BAKER_RATIO = 1F / 1F;
+	public static final float FARM_TO_MILL_RATIO = 3F / 1F;
+	public static final float FARM_TO_WATERWORKS_RATIO = 3F / 1F;
+	public static final float FARM_TO_PIG_FARM_RATIO = 3F / 1F;
+	public static final float FARM_TO_SLAUGHTER_RATIO = 6F / 1F;
+	public static final float WEAPON_SMITH_TO_LUMBERJACK_RATIO = 20F / 8F;
+	public static final float LUMBERJACK_TO_SAWMILL_RATIO = 2F / 1F;
+	public static final float LUMBERJACK_TO_FORESTER_RATIO = 2F / 1F;
+	public static final float LUMBERJACK_TO_STONE_CUTTER_RATIO = 8F / 5F;
+	public static final float COAL_MINE_TO_SMITH_RATIO = 1F / 1.8F;
+	public static final float IRON_MINE_TO_IRONMELT_RATIO = COAL_MINE_TO_SMITH_RATIO / COAL_MINE_TO_IRONORE_MINE_RATIO;
+	public static final float WINEGROWER_TO_TEMPLE_RATIO = 1F / 1F;
 }
